backend/pyproject.toml
[project]
name = "civics-platform"
version = "0.1.0"
description = "Civic Power Consortium - deliberation to allocation platform"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115",
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.14",
    "asyncpg>=0.30",
    "pydantic-settings>=2.6",
    "python-jose[cryptography]>=3.3",
    "httpx>=0.28",
    "sentry-sdk[fastapi]>=2.19",
    "structlog>=24.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
    "ruff>=0.8",
    "mypy>=1.13",
    "sqlalchemy[mypy]>=2.0",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["B008"]  # FastAPI Depends() pattern

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
